Refactored auth components, removed login component
