--- 03-Calling-an-API/src/app/login/login.component.html
@@ -1,1 +1,0 @@
-<app-loading></app-loading>
